+++ src/handlers/index.ts
@@ -4,3 +4,4 @@ export { handleUseTool } from "./useTool.js";
 export { handleHealthCheckAll } from "./healthCheck.js";
 export { handleAuthenticate } from "./authenticate.js";
 export { handleGetHelp, getTopicHelp, getErrorHelp, getPackageHelp } from "./getHelp.js";
+export { handleRestartPackage } from "./restartPackage.js";